+++ certbot/certbot/util.py
@@ -482,10 +482,11 @@ def enforce_le_validity(domain):
 
     """
     domain = enforce_domain_sanity(domain)
-    if not re.match("^[A-Za-z0-9.-]*$", domain):
+    if not re.match(r"^(\*\.)?[A-Za-z0-9.-]*$", domain):
         raise errors.ConfigurationError(
             "{0} contains an invalid character. "
-            "Valid characters are A-Z, a-z, 0-9, ., and -.".format(domain))
+            "Valid characters are A-Z, a-z, 0-9, ., and -. A wildcard label (*.) "
+            "may be present, but only as the first label.".format(domain))
 
     labels = domain.split(".")
     if len(labels) < 2: